Backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "unlabel-backend"
version = "0.1.0"
description = "Backend for Unlabel"
requires-python = ">=3.10"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "cachetools>=6.2.4",
    "certifi>=2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "dnspython==2.8.0",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "fastapi==0.128.0",
    "google-ai-generativelanguage==0.6.9",
    "google-api-core",
    "google-api-python-client",
    "google-auth",
    "google-auth-httplib2",
    "googleapis-common-protos",
    "grpcio>=1.60.0,<1.70.0",
    "grpcio-status>=1.60.0,<1.70.0",
    "h11==0.16.0",
    "httplib2",
    "idna==3.11",
    "proto-plus",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic_core==2.41.5",
    "pyparsing==3.3.1",
    "python-dotenv==1.2.1",
    "requests",
    "rsa==4.9.1",
    "six==1.17.0",
    "starlette==0.50.0",
    "tqdm==4.67.1",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "uritemplate",
    "urllib3>=2.6.2",
    "uvicorn==0.40.0",
    "google-generativeai==0.8.1",
    "python-multipart==0.0.21",
    "pillow>=12.0.0",
    "mangum>=0.17.0"
]